socketio 周りの整理
