+++ nodejs/socketIoServerModule/epgrecKeywordtable/deleteKeyword.js
@@ -1,0 +1,12 @@
+var log = require(__dirname + "/../..//logger").getLogger();
+var epgrecManager = require(__dirname + '/../../epgrecManager');
+
+module.exports = function(io, socket) {
+    //自動録画キーワード削除部分
+    socket.on("requestDeleteKeyword", function (id) {
+        epgrecManager.getDeleteKeywordResult(id, function(result) {
+                                            io.sockets.emit("resultDeleteKeyword", result);
+                                        });
+    });
+}
+
