+++ nodejs/socketIoServerModule/epgrecKeywordtable/setup.js
@@ -1,0 +1,6 @@
+var deleteKeyword = require(__dirname + '/deleteKeyword');
+
+module.exports = function(io, socket) {
+    deleteKeyword(io, socket); //自動録画キーワード削除部分
+}
+
