+++ nodejs/socketIoServerModule/epgrecRecorded/deleteVideoFile.js
@@ -1,0 +1,12 @@
+var log = require(__dirname + "/../..//logger").getLogger();
+var epgrecManager = require(__dirname + '/../../epgrecManager');
+
+module.exports = function(io, socket) {
+    //video file 削除部分
+    socket.on("requestDeleteVideoFile", function (rec_id, checkbox) {
+        epgrecManager.deleteVideoFile(rec_id, checkbox, function(result) {
+                                            io.sockets.emit("resultDeleteVideoFile", result);
+                                        });
+    });
+}
+
