+++ nodejs/socketIoServerModule/epgrecReservationtable/setup.js
@@ -1,0 +1,6 @@
+var cancelReservation = require(__dirname + '/cancelReservation');
+
+module.exports = function(io, socket) {
+    cancelReservation(io, socket); //録画予約一覧 削除部分
+}
+
